requirements.txt
streamlit
opencv-python-headless
numpy
pandas
altair
Pillow